+++ frontend/src/components/project/NewProject.tsx
@@ -4,6 +4,7 @@ import ExistingProjectForm from "./ExistingProjectForm";
 import NewProjectForm from "./NewProjectForm";
 import PageTitle from "../common/ui/PageTitle";
 import { Switch } from "../ui/switch";
+import CustomSwitch from "../common/ui/CustomSwitch";
 
 function NewProject() {
   const [isExistingProject, setIsExistingProject] = useState(false);
@@ -17,11 +18,10 @@ function NewProject() {
           Does your project already have an Github repository?
         </h2>
 
-        <Switch
+        <CustomSwitch
           id="existing-user"
           checked={isExistingProject}
           onCheckedChange={setIsExistingProject}
-          className="data-[state=checked]:bg-dark data-[state=unchecked]:bg-zinc-500 dark:data-[state=checked]:bg-light dark:data-[state=unchecked]:bg-zinc-500"
         />
       </div>
