+++ frontend/src/components/project/SyncedProject.tsx
@@ -1,0 +1,12 @@
+import { Project } from "@/types/Project";
+import { Input } from "../ui/input";
+
+function SyncedProject({ project }: { project: Project }) {
+  return (
+    <div>
+   syned project
+    </div>
+  );
+}
+
+export default SyncedProject;
